requirements.txt
matplotlib==3.10.5
numpy==2.3.2
pykan==0.2.8
scipy==1.16.1
torch==2.5.1
torch==2.6.0+cu124
